Split results by month and generate an xlsx
The requester wanted a different CSV file (or sheet in an xlsx file) for each month. This builds on the script that retrieves "Relevant Tweets Per User By Month" by splitting the results per month. Each user is present in each file, even if they did not post (NAs are introduced).